requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
confusable-homoglyphs==3.2.0
defusedxml==0.6.0
Django==3.0.8
django-allauth==0.42.0
django-crispy-forms==1.9.2
django-registration==3.1
django-rest-auth==0.9.5
django-webpack-loader==0.7.0
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.10
oauthlib==3.1.0
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
vue==0.0.1
whitenoise==5.1.0
